Use custom timeout handler in queue-proxy
The http.TimeoutHandler will buffer the response body in memory
until either the request completes or the request times out.

This works well for HTTP, but is a problem for HTTP2 and gRPC
streaming requests, where responses should be written as each
sub-request is processed.

This commit enforces the timeout by processing the request in a separate
goroutine and panicing with http.ErrAbortHandler if the timeout is
reached. The http.ErrAborthandler is a canary error used by the net/http
and x/net/http2 packages to gracefully end the connection without dumping
the stack trace.